chore: remove unnecessary example test style: merge all configs in main.js
